+++ griptape/tokenizers/openai_tokenizer.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 import logging
-from attrs import define
+from attrs import define, field, Factory
 import tiktoken
 from typing import Optional
 from griptape.tokenizers import BaseTokenizer
@@ -41,6 +41,13 @@ class OpenAiTokenizer(BaseTokenizer):
         "text-embedding-3-small",
         "text-embedding-3-large",
     ]
+
+    max_input_tokens: int = field(
+        kw_only=True, default=Factory(lambda self: self._default_max_input_tokens(), takes_self=True)
+    )
+    max_output_tokens: int = field(
+        kw_only=True, default=Factory(lambda self: self._default_max_output_tokens(), takes_self=True)
+    )
 
     @property
     def encoding(self) -> tiktoken.Encoding: